+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added `onStarted`, `onSucceeded`, `onFailed` and `onCompleted` callbacks for `Job`, `Step` and `Task`.
 - Added feature to specify the color of the message when logging out with `LogColor` and `ConsoleColor`.
+- Improved log message from framework.
 
 ### Destructive Changes